[device-minion-runtime] Introduce message_types.h with message types struct definitions

+++ device-minion-runtime/src/shared/include/device-common/log_levels.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-typedef uint64_t log_level_t;
+typedef uint8_t log_level_t;
 
 typedef enum {
     LOG_LEVEL_CRITICAL = 0,
